ai-service/requirements.txt
# BlancBleu — AI Service Requirements
# Transport sanitaire NON urgent — microservice Python local

# ── API Framework ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9       # Upload fichiers (PMT)

# ── OCR ───────────────────────────────────────────────────────────────────────
pytesseract==0.3.10            # Wrapper Python pour Tesseract OCR
Pillow==10.3.0                 # Manipulation d'images
pdf2image==1.17.0              # Conversion PDF → images pour OCR

# ── NLP / Extraction ──────────────────────────────────────────────────────────
spacy==3.7.4                   # NER pour noms, lieux
# Modèle français à installer séparément :
# python -m spacy download fr_core_news_sm

# ── Optimisation de tournée ───────────────────────────────────────────────────
ortools==9.10.4067             # Google OR-Tools (VRP)

# ── Validation des données ────────────────────────────────────────────────────
pydantic==2.7.1                # Validation schémas (inclus dans FastAPI)

# ── HTTP Client (calcul distances OSRM) ──────────────────────────────────────
httpx==0.27.0                  # Client HTTP async

# ── Utilitaires ───────────────────────────────────────────────────────────────
python-dateutil==2.9.0         # Parsing dates flexibles
